presentations/test-the-web-forward-preso/index.html

<!doctype html>  
<html lang="en">
	
	<head>
		<meta charset="utf-8">
		
		<title>Test the Web Forward</title>
        <meta name="apple-mobile-web-app-capable" content="yes" />
		<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
		
		<link href='http://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
		
		<link rel="stylesheet" href="css/reset.css">
		<link rel="stylesheet" href="css/main.css">
		<link rel="stylesheet" href="css/print.css" type="text/css" media="print">

		<link rel="stylesheet" href="lib/zenburn.css">
		<link rel="stylesheet" href="css/adobe.css">
	</head>
	
	<body>
		
		<div class="reveal">

			<!-- Used to fade in a background when a specific slide state is reached -->
			<div class="state-background"></div>
			
			<!-- Any section element inside of this container is displayed as a slide -->
			<div class="slides">
				<section>
					<h2>Test the Web Forward</h2>
					<img src="assets/images/W3C_logo.jpg">
					<script>
						// Delicously hacky. Look away.
						if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) ) document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
					</script>
				</section>
				
				<section>
					<h2>Reftest Overview</h2>
					<blockquote>
					<ul>
					    <li class="fragment">Two files: Test file + Reference file</li>
					    <li class="fragment">Test file uses the feature under test</li>
					    <li class="fragment">Reference file is an exact visual match without using feature under test</li>
						<li class="fragment">Self-describing (works both manually and with automation)</li>
						<li class="fragment"> Cross-browser & cross-platform</li>
					</ul>
					</blockquote>
				</section>
				
				<section>
    				<h2>Directory structure</h2>
    				<section>
    				    <br>
    				    
    				    <ul>
    				    <li>CSSWG Test Repository</li>
    				    <ul>
    				        <li>test</li>
    				        <ul>
    				            <li>contributors</li>
						        <ul>
						            <li>yourName</li>
						            <ul>
						                <li>incoming</li>
    						            <li>submitted</li>
						            </ul>
						        </ul>
						    </ul>
						</ul>
						</li>
					</section>
					<section>
					    <img src="assets/images/directoryStructure.png"/>
					</section>
    			</section>
    			<section>
    				<h2>Sample Test Case</h2>
    				<section>
    					<br><br>
    					<p>Basic test for the transform property described in the CSS3 Transforms spec using the translate() function</p>
    					<br>
    					<br>
    					<p>Spec Links:</p>
    					<p style="font-size: 16pt">
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#transform-property" target="top">http://dev.w3.org/csswg/css3-transforms/#transform-property</a>
    					    <br>
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					    </p>
					</section>
					<section>
					    <img src="assets/images/specAssertion.png">
					</section>
	            </section>
				
				<section data-state="screenshotPage">
    			    <h3>Sample files</h3>
    			    <p>Test</p>
    				<img src="assets/images/001a.png"/>
    			    <p>Reference</p>
    			    <img src="assets/images/001b.png"/>
    			</section>
			
				<section>
					<h3>Examine the code</h3>
					<section data-state="screenshotPage">
    				    <p>Test File</p>
    					<img src="assets/images/002a.png"/>
    				</section>

    				<section data-state="screenshotPage">
    				    <p>Reference File</p>
    				    <img src="assets/images/002b.png"/>
    				</section>
            	</section>
				
				<section>
				    <h3>Required Metadata: Title</h3>
				    <section data-state="screenshotPage">
					    <p>Test Title</p>
					    <img src="assets/images/003a.png"/>
				    </section>

				    <section data-state="screenshotPage">
					    <p>Reference Title</p>
					    <img src="assets/images/003b.png"/>
				    </section>
                </section>
                
                <section>
    			    <h3>Required Metadata: Author</h3>
    			    <section data-state="screenshotPage">
    				    <p>Test Author</p>
    				    <img src="assets/images/004a.png"/>
    			    </section>

    			    <section data-state="screenshotPage">
    				    <p>Reference Author</p>
    				    <img src="assets/images/004b.png"/>
    			    </section>
                </section>
                
                <section data-state="screenshotPage">
    			    <h3>Required Metadata: Spec Links</h3>
    			    <p>Test Spec Links</p>
    			    <img src="assets/images/005.png"/>
    			</section>

                <section data-state="screenshotPage">
    			    <h3>Required Metadata: Reference file</h3>
    			    <p>Test Reference File</p>
    			    <img src="assets/images/006.png"/>
    			</section>
    			
    			<section>
    			    <section data-state="screenshotPage">
    			        <h3>Required Metadata: Assertion</h3>
    			        <p>Test Assertion</p>
    			        <img src="assets/images/007.png"/>
    			    </section>
    			    <section>
    			        <img src="assets/images/specAssertion.png">
    			    </section>
    			</section>
    			
    			<section>
    			    <h3>The Test</h3>
    			    <section data-state="screenshotPage">
    				    <p>Test element with the transform applied</p>
    				    <img src="assets/images/008a.png"/>
    			    </section>

    			    <section data-state="screenshotPage">
    				    <p>Reference element</p>
    				    <img src="assets/images/008b.png"/>
    			    </section>
                </section>
                
                <section>
    			    <h3>Visible failure</h3>
    			    <section data-state="screenshotPage">
    			        <p>Element is only visible when the test fails</p>
    			        <img src="assets/images/009a.png"/>
    			    </section>
    			    <section data-state="screenshotPage">
    				    <p>Failing test example</p>
    				    <img src="assets/images/009b.png"/>
    			    </section>
    			</section>
    			
    			<section>
    			    <h3>Self Description</h3>
    			    <section data-state="screenshotPage">
    				    <p>Simple statement of how the page renders when the test passes</p>
    				    <img src="assets/images/010a.png"/>
    			    </section>

    			    <section data-state="screenshotPage">
    				    <p>Matching statement in the reference file</p>
    				    <img src="assets/images/010b.png"/>
    			    </section>
                </section>
                
                <section>
        		    <h3>Add the elements</h3>
        		    <section data-state="screenshotPage">
        			    <p>Add the red & green elements to the test page</p>
        			    <img src="assets/images/011a.png"/>
        		    </section>

        		    <section data-state="screenshotPage">
        			    <p>Add the matching green element to the reference file</p>
        			    <img src="assets/images/011b.png"/>
        		    </section>
                </section>
    			
				<!-- Begin JS TEST -->
				<section>
					<h2>JavaScript Test Overview</h2>
					<br>
					<ul>
					    <li>JavaScript should be used when Ref Tests are not practical</li>
					    <li>JavaScript tests have some advantages over Ref Tests
					        <ul>
					            <li>Robust</li>
					            <li>Flexible</li>
					            <li>Performant</li>
					            <li>Powerful</li>
					        </ul>
					    </li>
					    <li>Ref Tests have some advantages over JavaScript tests
					        <ul>
					            <li>Tests verify complete rendering</li>
					            <li>Easy to understand</li>
					            <li>Avoids introducing additional technologies (js)</li>
					        </ul>
					    </li>
					    <li>W3C provides a framework (testharness) to simplify and standardize test creation</li>
					</ul>
				</section>
				
				<section>
				    <section>
				    <h3>testharness</h3>
				        <p>testharness.js and testharnessreport.js</p>
				        <img src="assets/images/jsTest_th.png">
				        <br><br>
    				    <ul>
    				        <li>JavaScript framework that provides an API for making common test assertions</li>
    				        <li>Should be used by all JavaScript tests</li>
    				        <li>Path must be to /resources/... prior to pushing to the test repository</li>
    				        <li>An element must exist in test file with id="log" for test results</li>
    				    </ul>
    				    
				    </section>
				    
				    <section>
        			    <h2>The <span style="color:red">test</span> function</h2>
        			    <br>
        			    <ul>
        			        <li>Documented and implemented in testharness.js</li>
        			        <li>Function signature: <span style="color:red">test(test_function, name, properties)</span></li>
        			        <li>Used to define an individual test case in a file</li>
        			        <li>The function argument must be a function object, not a function call</li>
        			        <li>The return value of the function object passed to test() (argument 1) determines Pass/Fail</li>
        			        <li>The name passed to test() (argument 2) should be a descriptive test name that conveys what is being tested
        			            <ul>
        			                <li>Name will be displayed in the test results table</li>
        			            </ul>
        			        </li>
        			        <li>Properties (argument 3) is an object that overrides default test properties
        			            <ul>
        			                <li>timeout is the only currently recognized property</li>
        			            </ul>
        			        </li>
        			    </ul>
        		    </section>
        		    
        		    <section>
        			    <h2>testharness.js assert methods</h2>
        			    <br>
        			    <ul>
        			        <li>testharness.js provides a number of assert methods to use for testing
        			            <ul>
        			                <li>Full list is provided in the comment section of the testharness.js file</li>
        			                <li>Whenever possible, these methods should be used</li>
        			            </ul>
        			        </li>
        			        <li>All assert methods contain an optional "description" argument that is only output if the assertion fails
        			            <ul>
        			                <li>The description will be added to the default output from testharness</li>
        			                <li>The description should be used to provide additional debug information in the case of a failure</li>
        			            </ul>
        			        </li>
        			        <li>All asserts must be located in a test()
        			            <ul>
        			                <li>asserts outside won't be detected correctly by the harness and may cause a file to stop testing</li>
        			            </ul>
        			    </ul>
        			</section>
				</section>

                	<section>
        				<h2>Sample Test Case</h2>
        				<section>
        					<br><br>
        					<p>Basic test for the transform property described in the CSS3 Transforms spec using the translate() function</p>
        					<br>
        					<br>
        					<p>Spec Links:</p>
        					<p style="font-size: 16pt">
        					    <a href="http://dev.w3.org/csswg/css3-transforms/#transform-property" target="top">http://dev.w3.org/csswg/css3-transforms/#transform-property</a>
        					    <br>
        					    <a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
    					    </p>
    					</section>
    					<section>
    					    <img src="assets/images/specAssertion.png">
    					</section>
    	            </section>

				<section data-state="screenshotPage">
				    <h3>Examine the code</h3>
					<img src="assets/images/jsTest_full.png"/>
				</section>				

			    <section data-state="screenshotPage">
				    <h3>Required Metadata: Flags</h3>
				    <p>Flags</p>
				    <img src="assets/images/jsTest_flags.png"/>
			    </section>
			    
			    <section data-state="screenshotPage">
			        <h3>Optional: CSS</h3>
				    <p>Margins & Padding</p>
				    <img src="assets/images/jsTest_mp.png"/>
			    </section>

			    <section data-state="screenshotPage">
			        <h3>Self Description</h3>
				    <p>Positioned to simplify code</p>
				    <img src="assets/images/jsTest_res.png"/>
			    </section>
			    
			    <section data-state="screenshotPage">
				    <h3>Test Results Container</h3>
				    <p>Element with id="log"</p>
				    <img src="assets/images/jsTest_log.png"/>
			    </section>
 			    
			    <section data-state="screenshotPage">
				    <h3>User Defined Function</h3>
				    <p>JavaScript Function returns actual value</p>
				    <img src="assets/images/jsTest_func.png"/>
			    </section>

			    <section data-state="screenshotPage">
				    <h3>Test Case</h3>
				    <p>Test Function with assert_equals</p>
				    <img src="assets/images/jsTest_test.png"/>
				    <br><br><br>
				    <ul>
				        <li>assert_equals(actual, expected, description)
				            <ul>
				                <li>actual = checkElementPosition("myDiv")</li>
				                <li>expected = "100,100"</li>
				                <li>description = "myDiv should be at 100,100"</li>
				            </ul>
				        </li>
				        <br>				        
				        <li>test(test_function, name, properties)
				            <ul>
				                <li>test_function = assert_equals(checkElementPosition("myDiv"), "100,100", "myDiv should be at 100,100")</li>
				                <li>name = "Transform_Translate_100px_100px"</li>
				                <li>properties = <span style="font-style:italic">not specified</span></li>
				            </ul>
				        </li>
				    </ul>
				        
			    </section>   
                
    			<section>
    			    <section data-state="screenshotPage">
    				    <h3>Test Results</h3>
    				    <p>PASS</p>
    				    <img src="assets/images/jsTest_PASS.png"/>
    			    </section>
    			    
                    <section data-state="screenshotPage">
    				    <h3>Test Results</h3>
    				    <p>FAIL</p>
    				    <img src="assets/images/jsTest_FAIL.png"/>
    			    </section>    			    
                </section>                                                    
    			
    			<section>
    			    <h1>Q & A</h1>
    			</section>
		
			</div>

			<!-- The navigational controls UI -->
			<aside class="controls">
				<a class="left" href="#">&#x25C4;</a>
				<a class="right" href="#">&#x25BA;</a>
				<a class="up" href="#">&#x25B2;</a>
				<a class="down" href="#">&#x25BC;</a>
			</aside>

			<!-- Displays presentation progress, max value changes via JS to reflect # of slides -->
			<div class="progress"><span></span></div>
			
		</div>

		<!-- Optional libraries for code syntax highlighting and classList support in IE9 -->
		<script src="lib/highlight.js"></script>
		<script src="lib/classList.js"></script>
		
		<script src="js/reveal.js"></script>
		
		<script>
			// Parse the query string into a key/value object
			var query = {};
			location.search.replace( /[A-Z0-9]+?=(\w*)/gi, function(a) {
				query[ a.split( '=' ).shift() ] = a.split( '=' ).pop();
			} );

			// Fires when a slide with data-state=customevent is activated
			Reveal.addEventListener( 'customevent', function() {
				alert( '"customevent" has fired' );
			} );

			// Fires each time a new slide is activated
			Reveal.addEventListener( 'slidechanged', function( event ) {
				// event.indexh & event.indexv
			} );

			Reveal.initialize({
				// Display controls in the bottom right corner
				controls: true,

				// Display a presentation progress bar
				progress: true,

				// If true; each slide will be pushed to the browser history
				history: true,

				// Loops the presentation, defaults to false
				loop: false,

				// Flags if mouse wheel navigation should be enabled
				mouseWheel: true,

				// Apply a 3D roll to links on hover
				rollingLinks: true,

				// UI style
				theme: query.theme || 'default', // default/neon

				// Transition style
				transition: query.transition || 'default' // default/cube/page/concave/linear(2d)
			});

			hljs.initHighlightingOnLoad();
		</script>

	</body>
</html>
